chore: clean up clippy/dead-code lints and fix redis tls + mail features
Make the `--all-features` lints pass (clippy --workspace --all-targets
--all-features -D warnings is now clean) and fix feature-gated breakage
surfaced only under --all-features.

Fixes:
- armature-redis: `tls` feature now uses `redis/tokio-rustls-comp` (pulls
  tls-rustls AND the Tokio TLS impl); `redis/tls-rustls` alone left
  `connect_tcp_tls` unimplemented, breaking --all-features.
- armature-mail: add `base64` dep (sendgrid) and `multipart` reqwest feature
  (mailgun); add missing `TeraEngine` placeholder so the `tera` feature
  compiles (mirrors the minijinja stub).
- armature-gcp: build service clients without holding the parking_lot lock
  across `.await` (await_holding_lock) in all init_* functions.
- armature-core: prefix unused `_size` param in serialize_json_simd_with_size.
- armature-rhai: export ScriptContextBuilder/ScriptRouterBuilder/
  ConcurrentScriptCache/ScriptHandlerFn/script_handler/HotReloadExt (were
  defined but never re-exported -> dead_code).
- Misc clippy: needless borrows, redundant derefs, derivable impls,
  collapsible-if, unused imports (auto-fixed); targeted #[allow] for
  intentional dead code (diesel/graphql/websocket fields, bench structs),
  large_enum_variant (opensearch), wrong_self_convention (files), and the
  upstream-deprecated Zipkin exporter (opentelemetry).

+++ armature-diesel/src/pool.rs
@@ -155,6 +155,7 @@ impl PgPoolBb8 {
 #[cfg(all(feature = "mysql", feature = "deadpool"))]
 pub struct MysqlPool {
     pool: DeadpoolPool<AsyncMysqlConnection>,
+    #[allow(dead_code)] // kept for API symmetry with PgPool; not yet read on the mysql path
     config: Arc<DieselConfig>,
 }
 
@@ -196,7 +197,7 @@ impl MysqlPool {
         let status = self.pool.status();
         PoolStatus {
             size: status.size,
-            available: status.available as usize,
+            available: status.available,
             waiting: status.waiting,
             max_size: status.max_size,
         }
